Add files via upload
*Background info*
There is a new Excel file with highly sensitive data submitted regularly to a project team. The file and its sheets have their names changed by the data providers using a specific naming convention. The file's extra security restrictions prevent it from simply dropping its password in Excel and saving as an unprotected file. Other tools and macros would still see the file as protected, preventing it from further processing. The only working solution is to open a second instance of MS Excel and copy the restricted file's content with a formula. The script can be run from any directory multiple times as it overwrites an unprotected file if it already exists.